+++ app/config/config.php
@@ -1,2 +1,7 @@
 <?php
-require_once 'functions/functions.php';
+$db = [
+    'host' => 'localhost',
+    'user' => 'root',
+    'password' => 'root',
+    'database' => 'doingsdone'
+];
