+++ app/config/settings.php
@@ -1,0 +1,4 @@
+<?php
+require_once ('../functions/functions_db.php');
+
+$con = dbConnect($db);
