+++ app/index.php
@@ -1,6 +1,6 @@
 <?php
-require_once 'config/config.php';
 require_once 'functions/functions.php';
+require_once 'functions/functions_db.php';
 require_once 'helpers.php';
 
 $title = 'Дела в порядке: главная';
